+++ core/assignment.go
@@ -2,7 +2,22 @@ package core
 
 import (
 	"errors"
+	"fmt"
+	"math"
 	"math/big"
+)
+
+const (
+	percentMultiplier    = 100
+	minChunkLength       = 1
+	maxRequiredNumChunks = 8192
+)
+
+var (
+	ErrInvalidChunkLength  = errors.New("invalid chunk length")
+	ErrChunkLengthTooSmall = errors.New("chunk length too small")
+	ErrChunkLengthTooLarge = errors.New("chunk length too large")
+	ErrNotFound            = errors.New("not found")
 )
 
 // Assignment
@@ -33,146 +48,171 @@ func (c *Assignment) GetIndices() []ChunkNumber {
 	return indices
 }
 
+// Implementation
+
 // AssignmentCoordinator is responsible for taking the current OperatorState and the security requirements represented by a
 // given  QuorumResults and determining or validating system parameters that will satisfy these security requirements given the
 // OperatorStates. There are two classes of parameters that must be determined or validate: 1) the chunk indices that will be
 // assigned to each DA node, and 2) the size of each chunk.
-type AssignmentCoordinatorOld interface {
+type AssignmentCoordinator interface {
 
 	// GetAssignments calculates the full set of node assignments. The assignment of indices to nodes depends only on the OperatorState
 	// for a given quorum and the quantizationFactor. In particular, it does not depend on the security parameters.
-	GetAssignments(state *OperatorState, quorumID QuorumID, quantizationFactor uint) (map[OperatorID]Assignment, AssignmentInfo, error)
+	GetAssignments(state *OperatorState, blobLength uint, info *BlobQuorumInfo) (map[OperatorID]Assignment, AssignmentInfo, error)
 
 	// GetOperatorAssignment calculates the assignment for a specific DA node
-	GetOperatorAssignment(state *OperatorState, quorum QuorumID, quantizationFactor uint, id OperatorID) (Assignment, AssignmentInfo, error)
+	GetOperatorAssignment(state *OperatorState, header *BlobHeader, quorum QuorumID, id OperatorID) (Assignment, AssignmentInfo, error)
 
 	// GetMinimumChunkLength calculates the minimum chunkSize that is sufficient for a given blob for each quorum
-	GetMinimumChunkLength(numOperators, blobLength, quantizationFactor uint, quorumThreshold, adversaryThreshold uint8) (uint, error)
+	ValidateChunkLength(state *OperatorState, header *BlobHeader, quorum QuorumID) (bool, error)
 
-	// GetChunkLengthFromHeader calculates the chunk length from the blob header
-	GetChunkLengthFromHeader(state *OperatorState, header *BlobQuorumInfo) (uint, error)
+	CalculateChunkLength(state *OperatorState, blobLength uint, param *SecurityParam) (uint, error)
 }
 
-// Implementation
+type StdAssignmentCoordinator struct {
+}
 
-const PercentMultiplier = 100
+var _ AssignmentCoordinator = (*StdAssignmentCoordinator)(nil)
 
-var (
-	ErrNotFound = errors.New("not found")
-)
+func (c *StdAssignmentCoordinator) GetAssignments(state *OperatorState, blobLength uint, info *BlobQuorumInfo) (map[OperatorID]Assignment, AssignmentInfo, error) {
 
-// type StdAssignmentCoordinatorOld struct {
-// }
+	quorum := info.QuorumID
 
-// var _ AssignmentCoordinatorOld = (*StdAssignmentCoordinatorOld)(nil)
+	numOperators := len(state.Operators[quorum])
+	chunksByOperator := make([]uint, numOperators)
 
-// func (c *StdAssignmentCoordinatorOld) GetAssignments(state *OperatorState, quorum QuorumID, quantizationFactor uint) (map[OperatorID]Assignment, AssignmentInfo, error) {
+	// Get NumPar
+	numChunks := uint(0)
+	totalStakes := state.Totals[quorum].Stake
+	for _, r := range state.Operators[quorum] {
 
-// 	numOperators := len(state.Operators[quorum])
-// 	numOperatorsBig := new(big.Int).SetUint64(uint64(numOperators))
+		// m_i = ceil( B*S_i / C \gamma \sum_{j=1}^N S_j )
+		num := new(big.Int).Mul(big.NewInt(int64(blobLength*percentMultiplier)), r.Stake)
 
-// 	quantizationFactorBig := new(big.Int).SetUint64(uint64(quantizationFactor))
+		gammaChunkLength := big.NewInt(int64(info.ChunkLength) * int64((info.QuorumThreshold - info.AdversaryThreshold)))
+		denom := new(big.Int).Mul(gammaChunkLength, totalStakes)
+		m := roundUpDivideBig(num, denom)
 
-// 	chunksByOperator := make([]uint, numOperators)
+		numChunks += uint(m.Uint64())
+		chunksByOperator[r.Index] = uint(m.Uint64())
+	}
 
-// 	// Get NumPar
-// 	numChunks := uint(0)
-// 	totalStakes := state.Totals[quorum].Stake
-// 	for _, r := range state.Operators[quorum] {
+	currentIndex := uint(0)
+	assignments := make([]Assignment, numOperators)
+	for operatorInd := range chunksByOperator {
 
-// 		m := new(big.Int).Mul(numOperatorsBig, r.Stake)
-// 		m = m.Mul(m, quantizationFactorBig)
-// 		m = roundUpDivideBig(m, totalStakes)
+		// Find the operator that should be at index currentIndex
+		m := chunksByOperator[operatorInd]
+		assignments[operatorInd] = Assignment{
+			StartIndex: currentIndex,
+			NumChunks:  m,
+		}
+		currentIndex += m
+	}
 
-// 		numChunks += uint(m.Uint64())
-// 		chunksByOperator[r.Index] = uint(m.Uint64())
-// 	}
+	assignmentMap := make(map[OperatorID]Assignment)
 
-// 	currentIndex := uint(0)
-// 	assignments := make([]Assignment, numOperators)
+	for id, opInfo := range state.Operators[quorum] {
+		assignment := assignments[opInfo.Index]
+		assignmentMap[id] = assignment
+	}
 
-// 	headerHash := [32]byte{}
+	return assignmentMap, AssignmentInfo{
+		TotalChunks: numChunks,
+	}, nil
 
-// 	for orderedInd := range chunksByOperator {
+}
 
-// 		// Find the operator that should be at index currentIndex
-// 		operatorInd := getOperatorAtIndex(headerHash, orderedInd, numOperators)
-// 		m := chunksByOperator[operatorInd]
+func (c *StdAssignmentCoordinator) GetOperatorAssignment(state *OperatorState, header *BlobHeader, quorum QuorumID, id OperatorID) (Assignment, AssignmentInfo, error) {
 
-// 		assignments[operatorInd] = Assignment{
-// 			StartIndex: currentIndex,
-// 			NumChunks:  m,
-// 		}
-// 		currentIndex += m
-// 	}
+	assignments, info, err := c.GetAssignments(state, header.Length, header.QuorumInfos[quorum])
+	if err != nil {
+		return Assignment{}, AssignmentInfo{}, err
+	}
 
-// 	assignmentMap := make(map[OperatorID]Assignment)
+	assignment, ok := assignments[id]
+	if !ok {
+		return Assignment{}, AssignmentInfo{}, ErrNotFound
+	}
 
-// 	for id, opInfo := range state.Operators[quorum] {
-// 		assignment := assignments[opInfo.Index]
-// 		assignmentMap[id] = assignment
-// 	}
+	return assignment, info, nil
+}
 
-// 	return assignmentMap, AssignmentInfo{
-// 		TotalChunks: numChunks,
-// 	}, nil
+func (c *StdAssignmentCoordinator) ValidateChunkLength(state *OperatorState, header *BlobHeader, quorum QuorumID) (bool, error) {
 
-// }
+	info := header.QuorumInfos[quorum]
 
-// // getOperatorAtIndex returns the operator at a given index within the reordered sequence.
-// // We reorder the sequence by letting the reordered_index = operator_index + headerHash.
-// // Thus, get get the operator at a given reordered_index, we simply reverse:
-// // operator_index = reordered_index - headerHash
-// func getOperatorAtIndex(headerHash [32]byte, index, numOperators int) int {
-// 	indexBig := new(big.Int).SetUint64(uint64(index))
-// 	offset := new(big.Int).SetBytes(headerHash[:])
+	// Check that the chunk length meets the minimum requirement
+	if info.ChunkLength < minChunkLength {
+		return false, ErrChunkLengthTooSmall
+	}
 
-// 	operatorIndex := new(big.Int).Sub(indexBig, offset)
+	// Get minimum stake amont
+	minStake := state.Totals[quorum].Stake
+	for _, r := range state.Operators[quorum] {
+		if r.Stake.Cmp(minStake) < 0 {
+			minStake = r.Stake
+		}
+	}
 
-// 	operatorIndex.Mod(operatorIndex, new(big.Int).SetUint64(uint64(numOperators)))
+	totalStake := state.Totals[quorum].Stake
+	if info.ChunkLength != minChunkLength {
 
-// 	return int(operatorIndex.Uint64())
-// }
+		num := new(big.Int).Mul(big.NewInt(2*int64(header.Length*percentMultiplier)), minStake)
+		denom := new(big.Int).Mul(big.NewInt(int64(info.QuorumThreshold-info.AdversaryThreshold)), totalStake)
+		maxChunkLength := uint(roundUpDivideBig(num, denom).Uint64())
 
-// func (c *StdAssignmentCoordinatorOld) GetOperatorAssignment(state *OperatorState, quorum QuorumID, quantizationFactor uint, id OperatorID) (Assignment, AssignmentInfo, error) {
+		maxChunkLength2 := roundUpDivide(2*header.Length*percentMultiplier, maxRequiredNumChunks*uint(info.QuorumThreshold-info.AdversaryThreshold))
 
-// 	assignments, info, err := c.GetAssignments(state, quorum, quantizationFactor)
-// 	if err != nil {
-// 		return Assignment{}, AssignmentInfo{}, err
-// 	}
+		if maxChunkLength < maxChunkLength2 {
+			maxChunkLength = maxChunkLength2
+		}
 
-// 	assignment, ok := assignments[id]
-// 	if !ok {
-// 		return Assignment{}, AssignmentInfo{}, ErrNotFound
-// 	}
+		maxChunkLength = uint(nextPowerOf2(uint64(maxChunkLength)))
 
-// 	return assignment, info, nil
-// }
+		if info.ChunkLength > maxChunkLength {
+			fmt.Println("maxChunkLength", maxChunkLength, "info.ChunkLength", info.ChunkLength)
+			return false, ErrChunkLengthTooLarge
+		}
 
-// func (c *StdAssignmentCoordinatorOld) GetMinimumChunkLength(numOperators, blobLength, quantizationFactor uint, quorumThreshold, adversaryThreshold uint8) (uint, error) {
+	}
 
-// 	if adversaryThreshold >= quorumThreshold {
-// 		return 0, errors.New("invalid header: quorum threshold does not exceed adversary threshold")
-// 	}
+	return true, nil
 
-// 	numSys := roundUpDivide(uint(quorumThreshold-adversaryThreshold)*numOperators*quantizationFactor, PercentMultiplier)
-// 	chunkLength := roundUpDivide(blobLength, numSys)
-// 	return chunkLength, nil
+}
 
-// }
+func (c *StdAssignmentCoordinator) CalculateChunkLength(state *OperatorState, blobLength uint, param *SecurityParam) (uint, error) {
 
-// func (c *StdAssignmentCoordinatorOld) GetChunkLengthFromHeader(state *OperatorState, header *BlobQuorumInfo) (uint, error) {
+	quorum := param.QuorumID
+	numOperators := len(state.Operators[quorum])
 
-// 	// Validate the chunk length
-// 	numOperators := uint(len(state.Operators[header.QuorumID]))
-// 	chunkLength := header.EncodedBlobLength / (header.QuantizationFactor * numOperators)
+	// Get minimum stake amont
+	minStake := state.Totals[quorum].Stake
+	for _, r := range state.Operators[quorum] {
+		if r.Stake.Cmp(minStake) < 0 {
+			minStake = r.Stake
+		}
+	}
 
-// 	if chunkLength*header.QuantizationFactor*numOperators != header.EncodedBlobLength {
-// 		return 0, errors.New("invalid header")
-// 	}
+	totalStake := state.Totals[quorum].Stake
 
-// 	return chunkLength, nil
-// }
+	numChunks := int(roundUpDivideBig(totalStake, minStake).Uint64())
+
+	if numChunks > maxRequiredNumChunks-numOperators {
+		numChunks = maxRequiredNumChunks - numOperators
+	}
+
+	chunkLength := roundUpDivide(blobLength*percentMultiplier, uint(numChunks)*uint(param.QuorumThreshold-param.AdversaryThreshold))
+
+	if chunkLength < minChunkLength {
+		chunkLength = minChunkLength
+	}
+
+	chunkLength = uint(nextPowerOf2(uint64(chunkLength)))
+
+	return chunkLength, nil
+
+}
 
 func roundUpDivideBig(a, b *big.Int) *big.Int {
 
@@ -186,4 +226,9 @@ func roundUpDivideBig(a, b *big.Int) *big.Int {
 func roundUpDivide(a, b uint) uint {
 	return (a + b - 1) / b
 
+}
+
+func nextPowerOf2(d uint64) uint64 {
+	nextPower := math.Ceil(math.Log2(float64(d)))
+	return uint64(math.Pow(2.0, nextPower))
 }
