+++ disperser/batcher/encoding_streamer.go
@@ -580,42 +580,3 @@ func (e *EncodingStreamer) getOperatorStateForBlobs(ctx context.Context, metadat
 	return state, nil
 
 }
-
-// func (e *EncodingStreamer) getBatchMetadata(ctx context.Context, metadatas []*disperser.BlobMetadata, blockNumber uint) (*batchMetadata, error) {
-// 	quorums := make(map[core.QuorumID]QuorumInfo, 0)
-// 	for _, metadata := range metadatas {
-// 		for _, quorum := range metadata.RequestMetadata.SecurityParams {
-// 			quorums[quorum.QuorumID] = QuorumInfo{}
-// 		}
-// 	}
-
-// 	quorumIds := make([]core.QuorumID, len(quorums))
-// 	i := 0
-// 	for id := range quorums {
-// 		quorumIds[i] = id
-// 		i++
-// 	}
-
-// 	// Get the operator state
-// 	state, err := e.chainState.GetIndexedOperatorState(ctx, blockNumber, quorumIds)
-// 	if err != nil {
-// 		return nil, fmt.Errorf("error getting operator state at block number %d: %w", blockNumber, err)
-// 	}
-
-// 	for quorumID := range quorums {
-// 		assignments, info, err := e.assignmentCoordinator.GetAssignments(state.OperatorState, quorumID, QuantizationFactor)
-// 		if err != nil {
-// 			return nil, err
-// 		}
-// 		quorums[quorumID] = QuorumInfo{
-// 			Assignments:        assignments,
-// 			Info:               info,
-// 			QuantizationFactor: QuantizationFactor,
-// 		}
-// 	}
-
-// 	return &batchMetadata{
-// 		QuorumInfos: quorums,
-// 		State:       state,
-// 	}, nil
-// }
